==> 03_downstream_analysis/08_gene_importance/01_generatingSHAP_input.ipynb <==
import os

import scanpy as sc
import pandas as pd
import numpy as np

from glob import glob

from scipy.sparse import csr_matrix

from pyprojroot import here

### Loading data

**Selected genes from SPECTRA factors**

selectedGenes = pd.read_pickle(here('03_downstream_analysis/05_SPECTRA/results/SPECTRAFactor_selected_genes.pkl'))
selectedGenes

**ADATA object for each cell type**

adataPATHlist = glob(str(here('03_downstream_analysis/04_integration_with_annotation/results/normalized_adatas_nextflow/cellType_adata_merged/*_adataMerged.log1p.h5ad')))
len(adataPATHlist)

adataPATHlist = [p for p in adataPATHlist if not any(k in p for k in ['Progenitors','Cycling_cell'])]
len(adataPATHlist)

### Processing each object

Here we will 

1. load each merged object generated after get_normalized_expression with scANVI model.
2. subset object to keep only the 935 genes selected from SPECTRA factors
3. convert the result matrix into a sparse format
4. write the object for using it as input of Gradient Boosting Machine and SHAP analysis

for FP_i in adataPATHlist:
    ct_i = os.path.basename(FP_i).split('_adataMerged')[0]
    print(f"Reading {ct_i} ... ", end='')
    adata_i = sc.read_h5ad(FP_i, backed = 'r')
    print(f"done", end='\n')
    adata_i = adata_i[:,selectedGenes]
    print(f"Convert to sparse ... ", end='')
    adata_sparse_i = sc.AnnData(X = csr_matrix(adata_i.X), 
                                obs = adata_i.obs, 
                                var = adata_i.var, 
                                uns = adata_i.uns)
    print(f"done", end='\n')
    print(adata_sparse_i, end ='\n')
    adata_sparse_i.write(here(f'03_downstream_analysis/08_gene_importance/data/{ct_i}_adataMerged_SPECTRAgenes.log1p.h5ad'))
    print(f"{ct_i} adata wrote", end ='\n\n')
